+++ casdoorsdk/policy_global.go
@@ -1,0 +1,31 @@
+// Copyright 2024 The Casdoor Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package casdoorsdk
+
+func AddPolicy(enforcer *Enforcer, policy *CasbinRule) (bool, error) {
+	return globalClient.AddPolicy(enforcer, policy)
+}
+
+func UpdatePolicy(enforcer *Enforcer, oldpolicy *CasbinRule, newpolicy *CasbinRule) (bool, error) {
+	return globalClient.UpdatePolicy(enforcer, oldpolicy, newpolicy)
+}
+
+func RemovePolicy(enforcer *Enforcer, policy *CasbinRule) (bool, error) {
+	return globalClient.RemovePolicy(enforcer, policy)
+}
+
+func GetPolicies(enforcerName string, adapterId string) ([]*CasbinRule, error) {
+	return globalClient.GetPolicies(enforcerName, adapterId)
+}
